Change an exception to use failures
This adds a test to ensure we have objects of the right class thrown. The code
gets a bit bigger with this, and I am not sure the payload is neccesary, but I do
think adding the trace is a good idea. I added the payload mostly to show how it's
done for later.