Convert rest of exceptions to objects
Went through all `die` and `confess` calls I could find, and converted them to
failure->throw. The naming of exceptions is not consistent now, and I am struggling
to come up with good names, so if anyone else has ideas for what is best naming
policy, that would be great. I was struggling between naming for expected behaviour
or naming for what the error is, while trying to keep them short.

The test for `Tie::IxHash::Extension` throwing an exception object, is that the
code that uses it still survives and throws it's exepctions in the correct ways.

I uses `Safe::Isa` since that is what `failures` recommends.